pyproject.toml
[project]
name = "python-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.1",
    "langchain>=0.3.25",
    "openai>=1.82.1",
    "pgvector>=0.4.1",
    "polars>=1.30.0",
    "psycopg>=3.2.9",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "asyncpg-stubs>=0.30.1",
    "ruff>=0.11.12",
]
